+++ Assets/ConsentManagementProvider/Scripts/wrapper/iOS/ConsentWrapperIOS.cs
@@ -39,7 +39,9 @@ namespace ConsentManagementProviderLib.iOS
         [DllImport("__Internal")]
         private static extern void _configLib(int accountId, int propertyId, string propertyName, bool gdpr, bool ccpa, bool usnat, MESSAGE_LANGUAGE language, string gdprPmId, string ccpaPmId, string usnatPmId);
         [DllImport("__Internal")]
-        private static extern void _loadMessage(string authId);
+        private static extern void _loadMessage();
+        [DllImport("__Internal")]
+        private static extern void _loadMessageWithAuthId(string authId);
         [DllImport("__Internal")]
         private static extern void _loadGDPRPrivacyManager();
         [DllImport("__Internal")]
@@ -126,7 +128,16 @@ namespace ConsentManagementProviderLib.iOS
         public void LoadMessage(string authId = null)
         {
 #if UNITY_IOS && !UNITY_EDITOR_OSX
-            _loadMessage(authId);
+            if (string.IsNullOrEmpty(authId))
+            {
+                CmpDebugUtil.Log("Calling load message without authId");
+                _loadMessage();
+            }
+            else 
+            {
+                CmpDebugUtil.Log($"Calling load message with authId={authId}");
+                _loadMessageWithAuthId(authId);
+            }
 #endif
         }